Add constantSpeciesToParameters from AMICI. Start separate file for closely SBML-related functions (#58)
* Add constantSpeciesToParameters from AMICI. Start separate file for closely SBML-related functions

* Fix 'DeprecationWarning: The 'warn' method is deprecated, use 'warning' instead'

* Fix name

* Add test_constant_species_to_parameters

* Naming. Doc.
