+++ petab/__init__.py
@@ -1,3 +1,4 @@
 from .core import *  # noqa: F403, F401
 from .lint import *  # noqa: F403, F401
+from .sbml import *  # noqa: F403, F401
 from .version import __version__  # noqa: F401
